update backend map-generator to use faster golang-noise-module (fix #1)

+++ scripts/dev.sh
@@ -6,7 +6,18 @@ cd "$(dirname "$0")/../"
 
 PATH_FE="$(pwd)/src/frontend"
 PATH_BE="$(pwd)/src/backend"
+PATH_LIB="$(pwd)/lib"
 
+if ! [ -f "${PATH_LIB}/noise_cli" ] || ! [ -f "${PATH_LIB}/opensimplex_cli.py" ]
+then
+  echo "You need to install the noise-cli first: lib/README.md"
+  exit 1
+fi
+
+echo '### Running map-generator.. ###'
+python3 "${PATH_BE}/map_generator/main.py"
+
+echo '### Starting DEV Server... ###'
 trap 'kill $BGPID; exit' INT
 python3 scripts/dev_statics.py "$PATH_BE" &
 BGPID=$!
